Refactory API for organizing owner and member permissions

+++ app/Services/IdasService.php
@@ -11,6 +11,7 @@ namespace Figview\Services;
 
 use Figview\Repositories\IdasRepository;
 use Figview\Validators\IdasValidator;
+use LucaDegasperi\OAuth2Server\Facades\Authorizer;
 use Prettus\Validator\Exceptions\ValidatorException;
 
 class IdasService
@@ -51,6 +52,11 @@ class IdasService
 
     public function find($id)
     {
+        if($this->checkIdasOwner($id) == false)
+        {
+            return ['error' => 'Access Forbidden!'];
+        }
+
         return $this->repository->find($id);
     }
 
@@ -61,6 +67,11 @@ class IdasService
 
     public function update(array $data, $id)
     {
+        if($this->checkIdasOwner($id) == false)
+        {
+            return ['error' => 'Access Forbidden!'];
+        }
+        
         try{
             $this->validator->with($data)->passesOrFail();
             return $this->repository->update($data, $id);
@@ -75,7 +86,23 @@ class IdasService
 
     public function delete($id)
     {
+        if($this->checkIdasOwner($id) == false)
+        {
+            return ['error' => 'Access Forbidden!'];
+        }
         $this->repository->delete($id);
     }
 
+    /**
+     * Check if the user is owner of the resource.
+     *
+     * @param $idasId
+     * @return mixed
+     */
+    private function checkIdasOwner($idasId)
+    {
+        $userId = Authorizer::getResourceOwnerId();
+
+        return $this->repository->isOwner($idasId, $userId);
+    }
 }
