+++ app/Services/IoTEnvService.php
@@ -11,6 +11,7 @@ namespace Figview\Services;
 
 use Figview\Repositories\IotEnvRepository;
 use Figview\Validators\IotEnvValidator;
+use LucaDegasperi\OAuth2Server\Facades\Authorizer;
 use Prettus\Validator\Exceptions\ValidatorException;
 
 class IoTEnvService
@@ -50,6 +51,10 @@ class IoTEnvService
 
     public function find($id)
     {
+        if($this->checkIoTEnvPermissions($id) == false)
+        {
+            return ['error' => 'Access Forbidden!'];
+        }
         return $this->repository->find($id);
     }
 
@@ -60,6 +65,11 @@ class IoTEnvService
 
     public function update(array $data, $id)
     {
+        if($this->checkIoTEnvOwner($id) == false)
+        {
+            return ['error' => 'Access Forbidden!'];
+        }
+
         try{
             $this->validator->with($data)->passesOrFail();
             return $this->repository->update($data, $id);
@@ -74,6 +84,46 @@ class IoTEnvService
 
     public function delete($id)
     {
+        if($this->checkIoTEnvOwner($id) == false)
+        {
+            return ['error' => 'Access Forbidden!'];
+        }
         $this->repository->delete($id);
+    }
+
+    /**
+     * Check if the user is owner of the iotEnv.
+     *
+     * @param $iotEnvId
+     * @return mixed
+     */
+    private function checkIoTEnvOwner($iotEnvId)
+    {
+        $userId = Authorizer::getResourceOwnerId();
+
+        return $this->repository->isOwner($iotEnvId, $userId);
+    }
+
+    /**
+     * Check the user is member of the IoTEnv.
+     *
+     * @param $iotEnvId
+     * @return mixed
+     */
+    private function checkIoTEnvMember($iotEnvId)
+    {
+        $userId = Authorizer::getResourceOwnerId();
+
+        return $this->repository->hasMember($iotEnvId, $userId);
+    }
+
+    private function checkIoTEnvPermissions($iotEnvId)
+    {
+        if($this->checkIoTEnvOwner($iotEnvId) or $this->checkIoTEnvMember($iotEnvId))
+        {
+            return true;
+        }
+
+        return false;
     }
 }
